[mojo-lang] Implement initial support for `with` statements. (#13189)
This enables support for context managers in `with` statements.
These things are very helpful for implementing things like our
trace region support and things like Runtime support.

A context manager in Mojo implements three methods:

```
  fn __enter__(self) -> T:
  fn __exit__(self):
  fn __exit__(self, err: Error) -> Bool:
```

The first is invoked when the context is entered, and returns a
value that may optionally be bound to a target for use in the `with`
body.  If the `with` block exits normally, the second method is
invoked to clean it up.  If an error is raised, the third method
is invoked with the Error value.  If that method returns true, the
error is considered handled, if it returns false, the error is
rethrown so propagation continues out of the 'with' block.

This implementation has two significant limitations:
1) it only supports a single context mgr spec right now.
2) more significantly, it requires the enclosing context to be
   throwing.

I'll work to improve #2 before declaring success on this.

MODULAR_ORIG_COMMIT_REV_ID: 983e161ac63b7949c7f6f37c839c0994f30acbcb