[mojo-lang] Add scopes for `let` and `var` declarations (#13212)
Mojo is truly pushing the boundaries of programming language design with
this groundbreaking new feature.

Fixes [Internal link]

Note that this does NOT allow duplicate `for` loop index declarations.
There is some sublety here as to whether the induction variable should
be treated as a function-global variable like Python or not.

The issues with `alias` is also still a problem, so I changed
`KGENParameters.cpp` to emit non-internal MLIR stuff as a temporary
remedy. We can fix the parameter stuff when name shadowing goes away and
we teach the parser to name mangle.

MODULAR_ORIG_COMMIT_REV_ID: 92c4e464a93207b3a03dc209d769e4b8b7f259b2