[mojo-lldb] Handle the lldb events better (#13333)
There are two issues fixed by this change:
- Using the same listener for two different broadcasters results in some
events not being captured by the listener. This is probably a bug in
LLDB, but for now it's simpler to decouple this into two listeners.
- If the destructor of the REPL is invoked, the output of the last
expression might not be processed because of a race condition. A simple
way to fix this is to flush the output one more time after the event
listening loop has finished. This, along with
https://reviews.llvm.org/D149180, fixes the issues we've been having
with our tests.

MODULAR_ORIG_COMMIT_REV_ID: f5bbc87846c5f42066c675fe2c899410dbe0c063