[kernels] Improves the bandwidth when k is a multiple of 4096 with Graviton2 (#13318)
With Gravitron2 when k is a multiple of 4096 the FLOPS drops
significantly for matmul. This is likely due to addresses that differ by
the critical stride using the same cache line. One solution would be to
pad matrix A so that the stride is not a multiple of 4096. However, the
impact is much less with a 4x4 kernel than with the 8x2 kernel we
normally use so for now a 4x4 kernel is used when k is 4096 and neon
hardware is detected.

![image](https://user-images.githubusercontent.com/117054863/234397440-9fa8c974-7d4a-407b-96fd-746fc18036f2.png)

kernels-benchmark --filter=mojo.gemm.tiled --num-threads=1
Faux/utils/shapes/clipvit.yaml
```
shape             main    PR
257x4096x1024    35.431   35.529
257x1024x1024    35.483   35.702
257x1024x4096    24.080   31.226
```

Also the second shape in Faux/utils/shapes/top_models_ranges.yaml
```
shape             main     PR
256x1024x4096     24.203   31.330
```

LINALG_ORIG_REV_ID: e5d1bbd5f48f6e5f1698e2067e7ff3b8a218e3a5