[KGEN] For real fix the flaky elaborator failure (#13424)
I reverted the workaround but ended up hitting the error again locally.
The main issue is that we are somehow ending up with DTypeValues of the
same dtype but different data bitwidths, and this is non-determistic.
The culprit of the non-determinism is the attribute storage uniquer,
because the comparison only happens during the hash map lookup.

`SIMDType::readFrom` is correctly creating DTypeValue of bool dtype with
bitwidths of 8, but the DTypeValue constructor for bool is doing width
of 1. This inconsistency is causing the crash.

MODULAR_ORIG_COMMIT_REV_ID: 43a12874ffebddb20622047ae7dc7b2e0ad9fe7e