[KGEN] Expose `ObjectCompiler` functionality. (#13450)
This patch has two semi-hack changes. The first is that we expose the `isJIT` flag to the interface of the `lowerAllFuncsToLLVM` function. This is a hack because we shouldn't need this - it's only used in the REPL because we're forced into using the MCJIT for now. The second change is that we now allow users to update the pass manager used in the ObjectCompiler. This is necessary because the ObjectCompiler clears out the pass manager, and we would otherwise have to reset the whole pass pipeline. Given a good way to save and restore pass pipelines, we should do away with the second change as well.

MODULAR_ORIG_COMMIT_REV_ID: 5d1bbafbd9cdc464b8cb584c2657d2721481f972