[mojo-lldb] Fix another race condition in the tests (#13492)
For context, we have two threads: one that executes expressions and
another one that processes outputs and events. They don't have any sort
of synchronization point and sadly the REPL class doesn't offer any
place where we could do that, so it is possible that the events of
expression 1 are processed by the listener during of after the execution
of expression 2, which causes that the typesystem events and the stdout
of the inferior might get interleaved. This was causing some flakiness
in a test, where expressions are evaluated as fast as possible, as
opposed to a real user evaluating them one by one slowly and giving the
threads to catch up with each other. A simple way to address that is to
use FileCheck independently on the stdout and stderr streams.

Besides that I did some minor change in the listener code.

MODULAR_ORIG_COMMIT_REV_ID: b4933329852a304dfe1ecccf078a94cc12e108ad