[Mojo] Checking the full signature is incorrect for overloads. (#13498)
In order to merge a function into an overload set, we should check the
value inputs, parameter inputs, and function effects, not the full
signature. Basically, we currently consider return types when checking
if a function is an overload, and we currently don't handle that. We
could in the future, but for now, let's tighten the check to what we
actually support, which is overloading based on argument types/input
parameters.

Fixes #13484

MODULAR_ORIG_COMMIT_REV_ID: 8966ff371d1a56faf2371582af4a6b5b82b32e71