[mojo-lang] Reject attempts to define `__copyinit__/__moveinit__` as a def (#13527)
Defining the `__copyinit__` and `__moveinit__` functions for a type
as a def is fraught with peril: not only do you get implicit copies of
the incoming existing arguments in the copy case (which is doomed
to fail because it recursively invokes the copy ctor), this also makes
the special function implicitly throwing... which makes every copy/move
implicitly throwing, which isn't going to work out well.

Instead of allowing this, just punt on this with a clean error + fixit
that rewrites it to an 'fn'.  We aren't out to win religious wars here,
we are defining a pragmatic system, and 'fn' does the job we need done
here.

This fixes #13321

MODULAR_ORIG_COMMIT_REV_ID: 82bae01e76977594373484ef0f532f003727ce8e