[mojo-jupyter] Restrict the list of available LLDB commands (#13494)
We want to restrict the set of LLDB commands that can be executed on the
notebook as a way to prevent external users from affecting the host
environment (e.g. killing or spawning processes, inspecting the
entry-point, etc.). We do so by removing all commands except the few
that we include in an allowlist. We can grow this list as needed.

An effect of this is that the `help` command in fact shows only what is
available.

```
[0] > :help
[output] Debugger commands:
  apropos -- List debugger commands related to a word or subject.
  help    -- Show a list of all debugger commands, or give details about a
             specific command.
  log     -- Commands controlling LLDB internal logging.
Current user-defined container commands:
  mojo -- Commands related to the Mojo language support.
For more information on any command, type 'help <command-name>'.
```

MODULAR_ORIG_COMMIT_REV_ID: 145e24df63660735180a597e74e3525a4de47e0b