[mojo-lang] Add infra for body resolution decorators.  ~NFC (#13607)
This adds infrastructure to recognize that a decorator is not
handled at signature resolution time, so squirrel it away and
handle it when the body of the decl is resolved.  This moves
structs to this model (because `@value` decorator will need it),
but doesn't move functions or other things over yet.

MODULAR_ORIG_COMMIT_REV_ID: ed129f90040ed51de55a3b99ab76418fec29d966