[mojo-lang] Mangle invalid kwargs signature (#13675)
Currently, Mojo crashes when attempting to mangle a function with an
invalid signature (one that uses keyword arguments, which are not yet
supported):

```
fn g(*x: Int, y: Int): pass
```

```
error: TODO: keyword arguments not supported yet
fn g(*x: Int, y: Int): pass
              ^

Assertion failed: (isa<To>(Val) && "cast<Ty>() argument of incompatible type!"), function cast, file Casting.h, line 566.
```

The assertion occurs here:

```
frame #5 ASTType.cpp:199:18 → M::KGEN::LIT::ASTType::getVariadicElementType
   196  /// Given a VariadicType, return the element as an ASTType. This aborts if
   197  /// the current type isn't a VariadicType.
   198  ASTType ASTType::getVariadicElementType() const {
-> 199    return ASTType(cast<VariadicType>(mlirType).getElementType());
   200  }
   201
```

The problem is that `SignatureType::isVararg(size_t)` assumes that if
the argument index provided is the last one in the signature, and if the
signature is noted to have variadic arguments, then that last argument
must be the variadic one:

[Internal link]

But in this case, the signature is invalid (for now; once keyword
arguments are supported in Mojo, this will become a valid signature).
So, the above assumption doesn't hold, and `isVararg` returns the wrong
result.

For now, work around this problem by introducing (yet another) type
check, that the argument we're working with is of variadic type. This
isn't something we wish to keep doing, as described in #13015, but it's
a quick-and-dirty solution here.

Closes #13603.
MODULAR_ORIG_COMMIT_REV_ID: 1029a9d661262d7eaa50dc45c795541636da3bd6