[mojo-lang] Implement IRGen for capture-by-copy closures (Part 1) (#13646)
For launch, we will switch the default behaviour of nested functions to
always be value closures: non-parametric, always a dynamic value, and
only capture-by-copy.

This PR plumbs in a new bit on FuncOp, `nonparametric`, and implements
IRGen for capture-by-copy. The way this is implemented is a *little*
hacky, because we don't have the right model for it right now / are
lacking language features to implement a type-erased closure type in the
standard library.

When using a value captured from above inside a `@nonparametric`
closure, a copy is emitted above the closure the raw bytes are "moved"
into the closure by ensuring the bytes cross the closure boundary. This
rigs the closure lowering into generating a memcpy of the value into the
closure state, and ties the lifetime of the object with the closure.

Unfortunately, this will leak memory for now since closures don't have
destructors.

MODULAR_ORIG_COMMIT_REV_ID: 62bd7baad6ec02143a6831d52f31c15db4e3cce2