[mojo-lldb] Workaround for flaky tests (#13703)
For context, we have a race condition in the tests caused due to the
following:
The mojo plugin has a thread that produces at least three output streams
(stdout, stderr and one for other messages like diagnostics). However,
when LLDB finishes evaluating the expressions in the input test file, it
exits right away without waiting for the mojo plugin thread to process
these streams and output them for FileCheck to do its job.

I have an upstream fix that makes LLDB wait for that thread to finish on
exit, but as updating LLVM seems to be a lot of work this time, I'm
proposing this workaround which adds a sleep of 3 seconds at the end of
each test case so that there's enough time for the mojo plugin thread to
output the required streams.

MODULAR_ORIG_COMMIT_REV_ID: 94e0f60e0345a392f590d1ce6a516aa69a5ccfe3