[mojo-lldb] Use the correct lldb config when running tests (#13652)
For context, we've sporadically had timeout issues like the one in
[Internal link]

```
failed to launch REPL process: failed to get reply to handshake packet within timeout of 6.0 seconds
```

It turns out that this error is not unusual under if multiple heavy
tests are executed in parallel. The LLDB test suite addresses this issue
by setting the communication timeout to 60 secs in the lldb init file
`third-party/llvm-project/lldb/test/Shell/lit-lldb-init.in`. So in this
PR I'm making the lldb repl tests use the same configuration for
initialization.

For additional context, at some point I thought that the timeout was due
to ports not being available, but I was wrong: lldb by default connects
to the gdb-server using a socketpair, so that one shouldn't fail.

MODULAR_ORIG_COMMIT_REV_ID: 4c09457328e7f0b65d098d7814a919604c062b5f