[mojo-lang] Expand expression parser coverage. (#13857)
This expands and corrects the expression parser to be closer to
the python grammar, including support for the := operator (which
lacks codegen, but we parse it) and fixing representation of
starred item lists.  This is mostly a cleanup though.

MODULAR_ORIG_COMMIT_REV_ID: 891cbb416b7fcf3c65120bcb7da8307f0216bbf6