[KGEN] De-footgun the `getResolvedElementType` methods. (#14047)
The getResolvedElementType methods would check to see if the element
type is non-parametric and return the result if so, otherwise return
null.  This is a huge footgun, because clients want to be able to get
the element type of something even when parametric!

Fix this by wrapping in a ParamRefType, making this a total function.
Rename it to `getElementAsType()` to better clarify what it does.
This fixes #14018 and external issue 6.

MODULAR_ORIG_COMMIT_REV_ID: c2a6f355060954cb2e1a7db3bbc10c6597369c6f