[mojo-lang] Change by-ref syntax to inout syntax. (#14051)
This changes the syntax for by-refrence arguments to use the 'inout'
keyword instead of the `&` sigil.  This is more consistent with
borrowed/owned keywords, and is less confusing to new people.  The
term 'inout' is appropriate because computed lvalues need to copy
out the value (e.g. with getattr) then mutate, then copy in (with
setattr).

This moves a majority of the uses of & over, but not all of them.

LINALG_ORIG_REV_ID: ee7a1b9d9353d533cf1f543d2406537227fbc923