[Mojo][Doc-gen] Correctly handle ByRefReturns in throwing functions. (#14179)
When determining what result type to show in docs we need to take into account two things:
1. Whether the return value is passed by reference.
2. Whether the function throws (in that case the return type might be `Variant[Error, ActualReturnType]`)

(1) means that the return type is actually the type of the first argument of the function. (2) means that we want to extract `ActualReturnType` from the Variant type. However, we should only try to look into the variant type when the return value is not passed by reference - previously we didn't do it and it resulted in a crash on PythonInterop file.

MODULAR_ORIG_COMMIT_REV_ID: 8a2ff34a0c24f5e860004e8c7f3d14e9b6038a48