[mojo-repl] Simplify the event listening (#14213)
[mojo-repl] Simplify the event listening

Now that the debugger's default event thread launcher is public, we can
use it when starting the repl with `lldb -R`, which forced me to
implement an ugly one when we first started to work on this project.
When lldb is started normally, its command interpreter spawns that
thread on its own, unlike when using `lldb -R`. Therefore, we only need
to launch it ourselves when using `createInstanceFromDebugger` and not
when using `createInstanceFromTarget`, which is triggered by `expr -r`
in a normal debug session.

MODULAR_ORIG_COMMIT_REV_ID: 9c442dc46362667c6174feb380f82e7c152d5204