[mojo-jupyter] Highlight `borrowed|inout|owned` (#14281)
Currently `inout` is not highlighted in any particular way in Mojo
Jupyter notebooks, meaning it appears as ordinary black text on a white
background in `fn __init__(inout self, x: F32):`.

We could treat it as a keyword, in which case it would be highlighted in
the same way as `fn` or `def` (bold green). However, one user requested
it appear as an operator (bold purple), and so this pull request
highlights it in that way.

Do the same for `borrowed` and `owned`, too.

MODULAR_ORIG_COMMIT_REV_ID: a6d3b2c48a70d43da570ff5cc6b27b6812c78857