[mojo-lang] Enhance ternary expr to do conversions. (#14331)
When emitting a ternary operator with differing types on the true/false
values, we now see if exactly one is convertible to the other to form a
common type.  If so, we convert it and are good.  Otherwise we continue
rejecting if neither converts, and report an ambiguity if both convert.

This fixes [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: ecbf8476bb61c10cb7172cdbc5a27caa661426ed