[Mojo] Include Python Derived in Autoimports (#14370)
We need to make a python object printable. Without traits, this means we
need to add an initializer on the _Printable type with a PythonObject
which requires the Stdlib to have a dependency on Python. This can be
temporary. Once traits are done this can be reverted.

TODO: Add the reason this change was made to this PR so its tested.
MODULAR_ORIG_COMMIT_REV_ID: 9c13d4f372cfedf1958e7cb1bafa9cca2f95c34c