[KGEN] Lift FuncInterface Restriction on kgen.return (#14248)
Fixes [Internal link] . I'm fixing it
now because it is a prerequisite to
[Internal link] which is
a prerequisite for Python exceptions.

ReturnOp relies on the parent being a FunctionOpInterface in order to
call the function verifyReturnTypes. But really all it needs is the
expected return type of parent op. Ops like StageClosure and
CallThrowingInitOp have regions and should be able to contain return
values because they will eventually be lowered down to types that do
conform to FunctionOpInterface. But in the meantime its not necessary
for return op to require these ops to implement all the FunctionOp.

I introduced a new interface that allows function-like ops that have
regions but that are not yet functions (or inlined) to include return
ops in their regions. I did not require this though; rather I fall back
to FuncInterface if a parent with DefinesResult is not found. Reasons
are
[Internal link]

MODULAR_ORIG_COMMIT_REV_ID: d2a9be4edd4682cc75c8dd1d0d3b7b89e436b59b