[KGEN] Add `link` directive and add it to `external_call`. (#14516)
This patch adds 1 new op, `kgen.link`, and adds an optional attribute to `pop.external_call`. `pop.external_call` can now specify a symbol that its function is imported from, and that symbol is expected to be a KGEN::LinkOp. `kgen.link` specifies a library or object to be linked to the final archive, and the modifications to `pop.external_call` allow for `kgen.link` to be used.

This patch also adds the LLVM lowerings, which are roughly non-existent - `kgen.link` needs to survive until the `pop.external_call` is replaced, at which point it can simply disappear. The metadata it provides in the IR is used by the ObjectCompiler in `produceStandaloneArchive` before the LLVM lowering pipeline occurs.

MODULAR_ORIG_COMMIT_REV_ID: 02c98a93ae7fbc42fb49cf220498077efe138f3a