[KGEN] Add `-explicit-linking` and `-L` flags to `kgen` and `mojo`. (#14517)
This path adds flags that will be necessary in the future which allow the user to specify a link search path, as well as disabling the inclusion of CompilerRT in the current process. This is necessary because when we start statically linking CompilerRT, the JIT will complain about duplicate definitions.

MODULAR_ORIG_COMMIT_REV_ID: 1d66fead89f8841d0eb78acad820206e8dde2855