[LLCL] Guard TypeID name comparison in test (#14530)
The "Smoke" test case is failing when comparing the stringified type
name. This pretty type name is dependent on the host compiler and the
names are only guaranteed to be correct as written in the expectation
for clang right now. Codify that so the test won't fail on Windows when
using MSVC as the compiler, for example.

In the future, these names can be adjusted for MSVC if the Windows
bring-up folks care to test this.

Fixes [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 59588cfd4d98a62704b5295ddee542b6da598f1e