[mojo-lldb] Add commands for dynamically enabling and disabling LLVM debug logs (#13996)
These commands are

- llvm-debug enable [types]
- llvm-debug disable

They work for MojoLLDB plugin but not for MojoJupyter (the Kernel).
Sadly MojoJupyter has its own definition of the DebugFlag variable, so
there's no simple way to modify it from the plugin because the plugin is
a dependency of MojoJupyter. A possible workaround would be to move most
of the MojoJupyter code to MojoLLDB.

MODULAR_ORIG_COMMIT_REV_ID: a4b422b5c9c8c2ea81b74b3d27a728641ba237ea