[mojo-lldb] Extract to mojo only one declaration of each python variable (#14614)
Closes [Internal link]

Currently there's bug that happens when python code redefines a
variable, in which case we try to extract it to mojo twice, which
results in a compilation error. This also happens with imports.

A simple fix is to extract each variable only once.

A test is added.

MODULAR_ORIG_COMMIT_REV_ID: 0da97c5fcc95069dd575400dbbdb8eb87207ee7f