[KGEN] Make elaborator function name generation more deterministic (#14714)
The elaborator cannot rely on the internal counter in the symbol table
to generate deterministic names when it is parallelized.

It also makes it a bit nicer, mangling the fork value into the name and
removing the `_concrete` suffix.

MODULAR_ORIG_COMMIT_REV_ID: 5e8febf73fcac535468d817b5e732a56782f1776