[mojo-lldb] Return early if the python expression raises an exception (#14553)
If the python expression raises an exception, we are still trying to
extract the variables that it would have created, which fails and throws
with `Error: No module named '__lldb_python_module_*'`. As a way to
prevent this exception is simply to return early if the python
expression failed and thus skip the code that extracts these variables.

A test is added.

Closes [Internal link]
MODULAR_ORIG_COMMIT_REV_ID: d163eb41de41e863d44062d382ac5ccf895003df