[mojo-lang] Move SpecialFunctionKind up to FuncOp, NFC. (#14836)
This moves tracking of the special function kind (e.g. "are
you an __init__) up to the MLIR representation for a function
and out of ASTDecl.  This allows the optimizer to use it in
the future.  This is not being used yet though.

MODULAR_ORIG_COMMIT_REV_ID: e3ba4b56831195b8086c2a53728e85acf9a75242