[mojo][check-lifetimes] Implement copy->move optzn for reg-passable T's (#14838)
This enhances destructor insertion to notice if it is inserting a
destructor immediately after a copy constructor call, e.g. when about
to insert the del call in this pattern:

```
  %x = __copyinit__(%y)
  __del__(%y)
  ...
  use(%x)
```

when this happens, we can just remove the copy_init call and replace
uses of %x with %y.  This eliminates an extra copy of a value without
the user having the use the `x^` transfer operator explicitly.

This patch implements support for register_passable cases, memory
cases are slightly more involved so will go into another patch.

This is the first half of Issue #14504

MODULAR_ORIG_COMMIT_REV_ID: 6d728d8be6a81be0ee2aec29cacc88968cf29c03