[mojo][check-lifetimes] Implement copy->move optzn for memonly T's (#14840)
This extends the previous patch to elide temporaries in memory-only
calls when used in a limited idiomatic way.  This is enough to
reduce the need for explicit use of `^` operator as described in
the changelog.  This is part of Issue #14504.

There is a more general version of this that upgrades to moveinit,
but that is not included in this patch.

MODULAR_ORIG_COMMIT_REV_ID: 80120f4e0c55f2fd5776334019e8d06e85390777