[mojo-lang][Checklifetimes] Implement copyinit->moveinit promotion. (#14842)
This takes the next step in copyinit optimization in CheckLifetimes.
When we see a copy as the last use of a memoryonly type, but where
we cannot reuse the original input, we can still perform a move from
the input using the __moveinit__ function for the type.

This also extends the existing support for register passable types
to work more aggressively with vars and other indirectly stored
values.

The consequence of this is that PythonObject should be a lot more
efficient (among many other types) as mentioned in the release notes.
This is a big step beyond C++-like languages as well.

This completes Issue #14504.

MODULAR_ORIG_COMMIT_REV_ID: 054934f7025146724de1c31242d8bfbe35ff36db