[Stdlib] Renames the DType and aliases to be Int / Float instead of SI / F (#14863)
This renames the types to be more familiar to python programmers
(aligning mojo more closely to the numpy naming). Instead of exposing
the signess (which is more of a kgen detail) it exposes the types in
terms of intN, uintN, and floatN where N is the bitwidth.

Closes https://github.com/modularml/mojo/issues/152

KERNEL_BENCHMARKS_ORIG_REV_ID: 0d132b07201a8a1a4d10d56a739171e290f2e551