[API] Remove use of pyenv for Python  (#14618)
Now that we are more disciplined about our use of `virtualenv` when
doing Python things (including in CI), we don't need to use `pyenv` that is managed
by CMake for the API Python bindings.

Remove the pyenv entirely in favor of the virtualenv.  Add an option in
`setup-python` for installing some dependencies into the virtualenv for
executing the notebook tests that test the Python API for the Inference
Engine.

MAX_INTEGRATION_TESTS_REV_ID: f0d4f92e47f0939a58da4e93264b199248620fa4