[LLCL] Add an "awaiting" mutex (#14932)
This is a mutex that donates waiting threads to the runtime when waiting
on a mutex. This improves throughput in applications using LLCL because
threads waiting on mutexes for prolonged periods of time can still make
themselves useful. It is also required to implement a parallel
programming pattern in which a thread suspends inside a critical section
without deadlocking the program.

MODULAR_ORIG_COMMIT_REV_ID: 2c8a35659f4c65d1cc747507e112b6cf1c215251