[KGEN] Elaborate function forks in parallel (#14761)
This PR rewires the worklists in the elaborator to use AsyncValue and
an LLCL workqueue. Note that this PR implements very limited
parallelism:
only forks are handled in parallel. The next PR will substantially
increase parallelism by visiting callees in parallel, instead of
sequentially.

This PR also required introducing new state to each ParamNode that
involves flattening the callgraph. This is because the elaborator needs
to run cycle detection without relying on DFS order, and this is
unfortunate because there is no cheap multithreaded cycle detection
algorithm, especially one that will fit the unique circumstances of the
elaborator (the graph is not known ahead of time).

In addition, this PR ensures that evaluation is run, within a single
compiler process, is run in isolation by implementing a turnstile on the
evaluator function call that includes starving the LLCL workqueue so
that no other elaborator work is running at the same time.

MODULAR_ORIG_COMMIT_REV_ID: e5e300e0b7075b8ebcda509b5b04b1a27befffc1