[Cache][CAS] Make S3BlobCacheBackend subclass of BlobCacheBackend (#14880)
This makes the S3BlobCacheBackend a subclass of BlobCacheBackend, in
preparation to having the s3 backend redefine the insert/contains/find
methods (that will be a in follow-up PR).

NEW: I changed `M_CAS_allocateBackend` which has C linkage to return a
pointer to an incomplete type instead of LLCL::RCRef. Since it is a
pointer it shouldn't pose any problems with C. I wrap the pointer around
a RCRef in the dylib stub, because the backend is LLCL::ReferenceCounted
(due to being a subclass of BlobCacheBackend) and needs to be released
via the refcounting infrastructure.

MODULAR_ORIG_COMMIT_REV_ID: 0f10efff1bde0cf6ca47d7b7c6635b92afd01223