[LLCL] Another go at handling taskList overflow (#15059)
Though it would be great to allow the lock-free taskList to dynamically
grow when necessary, I can't figure out how to do that while still
retaining the independence of the push and pop operations.

So here's a middle ground: use an
old-fashioned-just-what-your-grandfather-wrote mutex guarded overflow
task list. For addTask, we only pay for the mutex on overflow. For
runItemsImpl, we only pay for the mutex when the worker would otherwise
go to sleep. Clearly tasks on the overflow queue can thus be starved, so
I added a fairness disclaimer to the API. But in return for this we can
now say addTask and addLocalTask NEVER run their closure immediately.

MODULAR_ORIG_COMMIT_REV_ID: 7f382b47e029898e99abea2c787d88968f2ee7d7