[KGEN][docs] Add a write-up on how parallel elaboration works (#15149)
Parallel elaboration got a lot hairier than anticipated, which is
something that was anticipated. There are plenty of comments in the
code, but none provide a high-level overview of how the parallelization
was achieved. This doc hopes to address that.

MODULAR_ORIG_COMMIT_REV_ID: dc0512c4e615b713020f8705ad50d5ac696646f1