[stdlib] Only execute async task inline if runtime has 1 thread (#15160)
[Internal link] made async_parallelize
execute the task in the host thread if there was only a single task to
run. The motivation for this was to reduce launch overhead in a very
tiny customer model executing on 1 thread.

However, it caused a ~2x regression on DLRM-rm2 by serializing the 40
embedding layers which take up majority of the runtime. The mef executor
launches the kernel from whichever thread emplaces the result that makes
the kernel ready to run. In RM2, the embedding layers all depend on a
single transpose, which means that they are all launched by the same
thread. Since they are small gathers, the kernels only use 1 thread, and
are then executed inline due to #14459. The result is that they are
serially executed in 1 thread while all the other threads are idle.

This change only runs the async_parallize task inline if the runtime has
a single thread, because that is the only case where it is gauranteed to
improve performance. In the future this heuristic should be extended to
handle other cases where it may be beneficial.

ALGORITHM_ORIG_REV_ID: 8c824937f78c3827938ec43e9f4a26cb107f0c0c