[mojo-lang] Improve diagnostic for raising outside an error context (#15221)
This error message was apparently confusing, and it was reported
multiple times. Improve the error with some hints. There were also cases
where the error is impossible because the raise is being emitted in a
context that is known to definitely raise already.

MODULAR_ORIG_COMMIT_REV_ID: 009808fc25879c899a59f59be3e41dc556c79746