[mojo-lang] Rename `TupleLiteral` to `Tuple`. (#15274)
As with Bool, this isn't a "literal type" that needs to convert to
other more concrete types.  This is a pervasive type that has sugar
(the comma syntax) that needs to be standardized across our
ecosystem.  Give it its rightful name! :-)

MODULAR_ORIG_COMMIT_REV_ID: 2a5272427603b913210f968245fcbb2246f5d9e3