[Stdlib] Implement gather in terms of elementwise generator (#15174)
The previous gather op was hardcoded to only work for specific shapes.
This lifts those constraints by providing a generic gather kernel that
is implemented in terms of the elementwise generator. The threading
logic moves from gather to the elementwise generator and this enables
elementwise fusions in the future.

In a followup, the constraints/reshaping in the shims will be removed as
well.

NN_ORIG_REV_ID: 1f5785bc31eb03e2b1c5f461efc9a5325f9a0a02