Revert "[KGEN] Disable region simplification in LLVM lowering pipeline" (#15297)
While this is a great idea, in practice this means that we are passing
more code to LLVM and causes longer build times. We are not going to
implement the sorting networks for now, so let's revert this.

This is a bandaid to tame the build times and we should figure out a
better solution in the G milestone.

Reverts [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 8e9a64a3a4c863b9ed157dc41a25c86134a44bee