[mojo-lang] Implement overloading on parameters (#15374)
This patch adds the ability to overload functions and methods on
parameters. The overload resolution logic prioritizes candidates with
the fewest implicit conversions in both the parameter and argument
signature. In case of a tie, the candidate without variadic arguments
and parameters is prioritized. If an unambiguous candidate cannot be
determined after this, signatures with the shortest input parameter
binding list are prioritized. These resolution rules are documented in
the programming manual.

The patch also modifies the MOGG binding resolution logic (and
corresponding tests), so that it handles mangled names correctly.

Closes [Internal link]
MODULAR_ORIG_COMMIT_REV_ID: 1e1fa73fd2e5e2e031980a2ef6937d94e8e74966