[mojo-driver] Attempt to un-flake mojo doc test (#15443)
We saw a stochastic failure in CI in:

[Internal link]

The output doesn't exactly show why this failed:

```
FAIL: KGEN :: mojo-driver/mojo-doc.mojo (65 of 179)
******************** TEST 'KGEN :: mojo-driver/mojo-doc.mojo' FAILED ********************
Script:
--
: 'RUN: at line 9';   echo "" | /__w/modular/modular/.derived/third-party/llvm-project/build-release/bin/not mojo-driver doc - 2>&1 | /__w/modular/modular/.derived/third-party/llvm-project/build-release/bin/FileCheck /__w/modular/modular/KGEN/test/mojo-driver/mojo-doc.mojo --check-prefix CHECK-NOT-MOJO
: 'RUN: at line 13';   not mojo-driver doc /does/not/exist.mojo 2>&1 | /__w/modular/modular/.derived/third-party/llvm-project/build-release/bin/FileCheck /__w/modular/modular/KGEN/test/mojo-driver/mojo-doc.mojo --check-prefix CHECK-BAD-INPUT
: 'RUN: at line 17';   not mojo-driver doc /__w/modular/modular/KGEN/test/mojo-driver/mojo-doc.mojo -o no/such/directory.mojo 2>&1 | /__w/modular/modular/.derived/third-party/llvm-project/build-release/bin/FileCheck /__w/modular/modular/KGEN/test/mojo-driver/mojo-doc.mojo --check-prefix CHECK-BAD-OUTPUT
: 'RUN: at line 21';   mojo-driver doc /__w/modular/modular/KGEN/test/mojo-driver/mojo-doc.mojo -o /dev/null
: 'RUN: at line 25';   mojo-driver doc /__w/modular/modular/KGEN/test/mojo-driver/mojo-doc.mojo -I /__w/modular/modular/KGEN/test/mojo-driver/mojo-doc.mojo -I /does/not/exist
--
Exit Code: 141
```

We don't know which `RUN` line failed, but the exit code of 141 points
to `141 - 128 = 13`, which, according to
https://www.man7.org/linux/man-pages/man7/signal.7.html, is a `SIGPIPE`
signal termination due to "Broken pipe: write to pipe with no readers."

My shot-in-the-dark theory is that writing to stdout (the default for
`mojo-driver doc`) can fail when run in CI. This commit instead writes
to a temporary file.

Also, instead of reading the `mojo-doc.mojo` file as both an input file
and an include, read another file as an include in this test.

MODULAR_ORIG_COMMIT_REV_ID: c1bb8c8637d89550237926e636bd12b4292e3155