[CAS] S3 backend: Use AWS async API and threads for downloads (#15070)
S3 throughput can be greatly increased by doing many transfers in
parallel (of different objects or different chunks of the same object).

This PR changes how the S3 backend downloads objects, by using the AWS
SDK async API for get object requests. What this means is that get
object requests are handled by AWS threads and we can download up to N
objects in parallel, where N is the size of the AWS thread pool. Reasons
for doing it this way:
- We can have N parallel downloads, where N is the number of IO threads.
- Using AWS threads for IO, we are not limited by the number of LLCL
threads. The number of LLCL threads is optimized for compute, not IO.
- We don't block LLCL threads on IO. IO takes time but uses very little
CPU, and by doing it in different threads we free up LLCL threads for
compute. Maybe this doesn't matter much now, but if in the future IO
also happens during computation, this would allow for more overlap of
communication and computation when using S3.
- AWS SDK has an async API and so implementing this is relatively
straightforward.

As an example, bert-large with externalized weights in S3 is a total of
around 1.4 GB and 400 files, of which some are very small. On
c5.4xlarge, doing initModel when the data is in S3 is ~50% faster when
using 16 IO threads instead of 8. Note that 8 is the default number of
LLCL threads in this machine (it has 8 physical cores). On c5.4xlarge,
bert-large initModel with this PR and using 16 AWS threads takes about 2
secs in the best case.

This PR only changes the implementation of find. I think insert and
contains are less important and can be done later.

MODULAR_ORIG_COMMIT_REV_ID: 3c5b088f6d563e555378b98b64cc716235f4ca33