[KGEN] Split out a `GeneratorUserOpInterface` from `KGENCallOpInterface` (NFC) (#15456)
`KGENCallOpInterface` was a combination of two things: an interface to
represent direct calls, and an interface to represent parametric
generator references for the elaborator. However, the new `kgen.cost_of`
operation is a generator user but not a call operation in the sense that
it has call arguments, call results, and could be inlined. This PR
splits a new `GeneratorUserOpInterface` out that is entirely an
elaborator interface, whereas the leftover `KGENCallOpInterface` is
purely for calls.

MODULAR_ORIG_COMMIT_REV_ID: 4c8b96d0a275f50fdef4d7b7e9a07f4edeab7b4a