[KGEN] Use a slab allocator for managing JIT memory (#15435)
This is the recommended memory management scheme in orc,
and on windows helps to ensure that code is allocated to allow
for 32 bit relocations. As part of this, remove the large code
model for COFF because that actually didn't help the relocation
issue we were seeing.

MODULAR_ORIG_COMMIT_REV_ID: da270e24d815f751332bd5d714b0c4cef3ed6c0c