[mojo-lldb] Fix a python expression invocation in a test (#15548)
This expression was actually starting with the comment in line 31, and
because of that it was being handled as mojo code. It was showing a
failure, but we weren't checking for its success anyway. Adding a blank
line fixes it.

MODULAR_ORIG_COMMIT_REV_ID: dd643ace5e706c51f4769179e32f4d62870d038f