[mojo-parser] Ignore by-ref results when looking for overlapping functions
Closes [Internal link]

When resolving overlapping functions in the following repeated cell

```
struct Struct:
    fn __init__(inout self):
        pass
fn xoo() -> Struct:
    return Struct()
let abc = xoo()
```

we end up seeing the following signatures for the first and second
execution of the cell

```
!kgen.signature<(!pop.pointer<@"$Expression [0]"::@Struct> byref_result) -> !lit.none>
!kgen.signature<(!pop.pointer<@"$Expression [1]"::@Struct> byref_result) -> !lit.none>
```

and when the parser was comparing their inputs to see if they match, it
wasn't ignoring the byref_result, so it thought that both the
expressions should be included when resolving function calls, which
terminated in failure. Only the second one should have been kept.

A simple fix suggested by River is to ignore the byref result in that
comparison.

MODULAR_ORIG_COMMIT_REV_ID: 26d57b04b0ae6870214b394ec7246a60fe766c1f