[KGEN] Explicitly link to libKGENCompilerRT within the JIT (#15495)
Right now we are in this quite awkward situation where we expose
CompilerRT symbols from the parent process, creating quite a few
issues:

* It makes it difficult to do any form of out-of-process/out-of-machine linking
* We need to manually register all of the CompilerRT functions with the JIT in
  every tool that uses the JIT.
* Windows JITTing doesn't like in process symbols, and has lots of issues linking
  to CompilerRT this way.

This PR seeks to remedy some of this by generating a shared library variant of
CompilerRT for use by the JIT. The JIT will load this dynamic library as a
generator, which fits better with the modeling of symbol resolution in the jit
(and also makes windows much happier). Eventually this linking should be
specified by the standard library on the mojo (either via link directives or
some other build specification). With this we can sever effectively all of the
tool dependencies on CompilerRT.

MODULAR_ORIG_COMMIT_REV_ID: cfd2ec497f45591bfe5856dff82aec0bc94b5992