[KGEN] Remove unused loop results (and fix dead loop removal) (#15603)
This PR adds a canonicalization pattern to remove unused `hlcf.loop`
results. This is added as a pattern and not to SimplifyCF because the
cost to traverse the loop body is paid only when the pattern matchers.

This PR also fixes some issues with the dead loop elimination: crashes
and invalid IR. Apparently one of the cases in that pattern was simply
never exercised.

MODULAR_ORIG_COMMIT_REV_ID: 58dbef6f099e8df8d235063d1e1ad8eab247bd60