[MOGG] Propagate user provided paramters through compiler (#15584)
This allows the compiler to set user provided parameters such as
"transpose_B" on matmul. These allow compiler interopt driven by the
mojo kernel. This is not the final solution of this problem. With the
rest of the interopt we will need to make this less handcoded but
atleast for now it can allow parameters to be passed through.

I started with single_thread_blocking_override which doesn't do anything
now but will allow the kernel to tell the compiler it can be toggled
into a blocking mode which allows us to compose soft fusion through
setting this toggle.

See the notion doc for more info on this flag:
[Internal link]

MOGG_ORIG_REV_ID: 0109e8651197ba55e72156c347d555b17d8f2120