[KGEN] Pass version explicitly for getting *RT from the CAS. (#15702)
This patch passes in CAS backend versions for storing and retrieving CompilerRT and OrcRT in the CAS. This ensures that the versions are the same - it does NOT make an attempt to find the 'best' version to use for this. It should fix the issue we are having where it's sometimes impossible to find the CompilerRT after a commit, even if the binary itself has not changed.

MODULAR_ORIG_COMMIT_REV_ID: 9dc13bbf494502d921bb02e89bb2ccb851b37a34