[Kernels] Make batched matmul jit'able by pulling it out of the shims (#15730)
Introduces a new function in LinAlg called batched_matmul_parallel_async which
does batched matmul on a NDBuffer.

Note that the shim reshapes the input buffers to 3D before calling the function
to avoid an explosion in the number of kernels that need to be instantiated.

Also a few minor changes:
- disable the adj_a flag for simplicity since it is implemented via a naive
matmul and there isn't a fusion pattern for it so it can't be called by the
graph compiler anyways
- explictly assert that max rank is 4 (will fix this and add tests in a followup)

LINALG_ORIG_REV_ID: e473cef2419d31ee7f0d2d6a30be510ad7f04686