[REPL] Handle correctly setups without lldb-python integration (#15735)
When the lldb-python integration is not enabled, 3 tests fail, so we are
now only running them if the integration is enabled. We are also forcing
the execution of these tests when in the CI.
Besides that, we were incorrectly checking if the integration was
enabled inside the LLDB plugin. Fixing that ensures that a warning
message is displayed to the user, so that they know why some features
don't work as normally expected.

Closes [Internal link]
Closes [Internal link]
Closes [Internal link]
MODULAR_ORIG_COMMIT_REV_ID: b8f0071e2270fbdb0e478b597a280f762e77588d