[KGEN] Stack reuse optimization pass (#15899)
This PR implements `stack-reuse`, an optimization pass that removes
redundant stack allocations while perfoming some copy elision. The jist
of the pass is that if at all points the projection of stack
allocation A (the stack-allocated pointer and all views of it, e.g. via
`gep`) is accessed via a `pop.load`, another stack allocation that
dominates A contains the same value, the A can be elided from the
program.

Enabling this pass reduces the size of `kgen -elaborate Misc.mojo` from
540040
lines of IR to 471100. That's 68940 ops removed!

MODULAR_ORIG_COMMIT_REV_ID: bb93e8240641257544620dfab1227fea81f4a207