[repl] Allow trailing spaces in the python magic line (#15835)
A few times I copy pasted python expressions into the repl from other
terminals, and that added trailing spaces in the %%python magic line,
which caused a weird error because we were executing the entire thing as
a mojo expression. That is very confusing, so I'm relaxing the check to
allow trailing spaces in the magic line. I also changed a little bit the
ordering of the conditions in that part of the code so that it's easier
to add more magic lines.

MODULAR_ORIG_COMMIT_REV_ID: 641629526a4578d70dec717e1de3288997cf4ad8