[KGEN] Fix calling of a macro in `MojoLLDB` (#16007)
On a compiler setup I'm playing with (LLVM from homebrew
with LLVM assertions off), this causes a hard error.

Fix the use of `setCurrentDebugTypes` macro to not have it be
namespace-qualified since macros are not namespace-aware. Note that this
entity is either a function in debug mode, or a macro otherwise. That's a bit
of a weird design in LLVM, but this fixes the hard error.

MODULAR_ORIG_COMMIT_REV_ID: a51a351177afbea0177b930d593a4b9809951000